Update parser.py to begin detecting missing sequencing ';' tokens
Update parser.py to begin detecting missing sequencing ';' tokens. This is still under development, but this update allows the parser to detect missing sequencing ';' (i.e. semicolons) at the top level and report a syntax error associated with premature ending of the parsing process (because without a semicolon, the next statement will not be parsed and we end up with unparsed tokens).