Update parser.py to consolidate some method code
Update parser.py to consolidate some method code into a separate statement_seq() method to deal with sequences of statements. This is then used not only at the top level (since the program itself is a sequence of statements) but also by the methods that parse if-then-else statements and while-do statements, since each of those also contain sequences of statements.