Reorganize parser.py for upgraded PT and AST
Reorganize and redevelop parser.py to make the parse tree (PT) more technically correct and to generate the PT in parallel with the generation of the abstract syntax tree (AST) instead of deriving the AST from the PT later. The PT is still not absolutely correct, fudging a bit at the top level of a sequence and using a flattened list instead of sequence of nested tuples at that level. This commit still keeps the code formerly used to transform the PT into the AST in case it's still useful or we discover we need to return to the previous method. This reorganization also requires some minor changes in the compiler.py file (see later commit).