Update parser.py to upgrade parse tree and maintain AST
Update parser.py to upgrade parse tree and maintain AST, with both still be generated in parallel (instead of my initial plan to reconstitute the code to generate the PT first and then derive the AST). The parallel approach still seems a bit easier than deriving the AST as a further-processed version of the PT, and now both seem to be working as desired. Some more testing is called for before concluding it works perfectly. This commit also deletes some obsolete code for PT-to-AST translation, which never quite worked anyway.