Update parser to produce Node-based parse tree
Update Parser class in parser.py to produce a TreeNode-based parse tree. The parser still also produces the nested-tuple string versions of both parse tree and AST, but will be editing that out next. The .dot versions of the TreeNode-based parse tree and AST all look good when tested on a variety of test WHILE programs.