Add __repr__() fxn to CFG class
Add __repr__() fxn to CFG class to facilitate printing and debugging. Played with various forms here, but eventually settled on a form that prints info for a single node at a time rather than recursively printing all successor nodes, in part because a CFG is not a tree and can lead to problematic recursion unless we take special precautions.