Modify node_from_cfg() to incorporate more of original AST
Modify node_from_cfg() to incorporate more of original AST upon which the CFG is based. In particular, augment AST component being saved in the IF and WHILE CFG nodes. This provides a mechanism in particular for recreating the structure of the if-then-else-fi constructs and better allows the generation of C code from the CFG.