Update optimizer.py to again deal with some SKIP stmts
Update optimizer.py to again deal with some SKIP statements, this time making some more nuanced distinctions among the various scenarios. The code can now remove "skip" nodes from the CFG in most cases, EXCEPT the case where the "skip" appears as the first element in the THEN block of an if-then-else-fi block.