Update optimizer.py and cfg.py to better deal with skip stmts
Update optimizer.py and cfg.py to better deal with skip stmts, basically removing the special-case approach I was trying in the optimizer in the previous commit and making a small modification in the cfg.py remove_node() method. The remove_node() method was modified to preserve the order of successors in the list of successors when a node removal occurs and leads to a successor being replaced with a subsequent (i.e. “grandchild” instead of child) successor. This code has been tested on a variety of while programs on the risc-v machine.